apps/studio/e2e_test_apps/app/pyproject.toml
[project]
name = "e2e-test-app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "junjo",
    "lorem-text>=3.0",
    "pyyaml",
]
